Further splitting of LeetCode API helper functions into multiple files
